+++ lib/bcdd/result/contract/for_types_and_values.rb
@@ -30,6 +30,8 @@ class BCDD::Result
     def type_and_value!(data)
       type, value = data.type, data.value
 
+      return value if IgnoredTypes.include?(type)
+
       value_checking = @types_and_values[type!(type)]
 
       checking_result = value_checking === value